requirements.txt
astroid==1.4.9
decorator==4.0.11
ipython-genutils==0.2.0
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.6.1
olefile==0.44
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.13
ptyprocess==0.5.1
Pygments==2.2.0
pylint==1.6.5
selenium==3.3.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
wrapt==1.10.10
